add tests & rearrange
